feat: add mouse controls + some minor reorganization
